refactor: décomposer Index.tsx en composants de pages (issue #374)
Extrait 4 pages du god component Index.tsx (2587 lignes) vers des
composants dédiés dans src/pages/game/ :

- SummonPage.tsx : Page 0 — Invoquer (onglets BC/Shards, dernière invocation)
- HeroesPage.tsx : Page 1 — Héros (collection, codex, équipes, filtres)
- ProgressionPage.tsx : Page 3 — Progression (stats, quêtes journalières, succès)
- ForgePage.tsx : Page 4 — Forge (fusion + recyclage)

Stratégie conservatrice : tout l'état (useState, useEffect, useCallback)
reste dans Index.tsx. Les composants reçoivent leurs données via props.
Page 2 (Combat + game loop) non extraite intentionnellement.

Index.tsx : 2587 → 1751 lignes (-836 lignes, -32%)
Nettoyage des imports non utilisés après extraction.

246 tests passent. Build OK.

Closes #374
